[Ref]  Extracting PhoneticSwapInternalState
Refactoring - Extracting `PhoneticSwapInternalState` to an independent file
